# Golf compile times