--- a/.changeset/hungry-spiders-beg.md
+++ b/.changeset/hungry-spiders-beg.md
@@ -1,0 +1,5 @@
+---
+"ccip": patch
+---
+
+golf compile times of tests