--- a/contracts/src/v0.8/ccip/test/StructFactory.sol
+++ b/contracts/src/v0.8/ccip/test/StructFactory.sol
@@ -1,17 +1,10 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity 0.8.24;
 
-import {IPool} from "../interfaces/IPool.sol";
-import {IPriceRegistry} from "../interfaces/IPriceRegistry.sol";
-
-import {PriceRegistry} from "../PriceRegistry.sol";
-import {RMN} from "../RMN.sol";
+// Imports to any onramp/commitStore/offramp are not allowed due to the significant cascading
+// compile time increase they cause when imported into this base test.
 import {Internal} from "../libraries/Internal.sol";
 import {RateLimiter} from "../libraries/RateLimiter.sol";
-
-import {EVM2EVMMultiOffRamp} from "../offRamp/EVM2EVMMultiOffRamp.sol";
-import {EVM2EVMOffRamp} from "../offRamp/EVM2EVMOffRamp.sol";
-import {EVM2EVMOnRamp} from "../onRamp/EVM2EVMOnRamp.sol";
 
 contract StructFactory {
   // Addresses
@@ -20,67 +13,11 @@ contract StructFactory {
   address internal constant DUMMY_CONTRACT_ADDRESS = 0x1111111111111111111111111111111111111112;
   address internal constant ON_RAMP_ADDRESS = 0x11118e64e1FB0c487f25dD6D3601FF6aF8d32E4e;
   address internal constant ZERO_ADDRESS = address(0);
-  address internal constant BLESS_VOTER_1 = address(1);
-  address internal constant CURSE_VOTER_1 = address(10);
-  address internal constant CURSE_UNVOTER_1 = address(110);
-  address internal constant BLESS_VOTER_2 = address(2);
-  address internal constant CURSE_VOTER_2 = address(12);
-  address internal constant CURSE_UNVOTER_2 = address(112);
-  address internal constant BLESS_VOTER_3 = address(3);
-  address internal constant CURSE_VOTER_3 = address(13);
-  address internal constant CURSE_UNVOTER_3 = address(113);
-  address internal constant BLESS_VOTER_4 = address(4);
-  address internal constant CURSE_VOTER_4 = address(14);
-  address internal constant CURSE_UNVOTER_4 = address(114);
 
   address internal constant USER_1 = address(1);
   address internal constant USER_2 = address(2);
   address internal constant USER_3 = address(3);
   address internal constant USER_4 = address(4);
-
-  // RMN
-  function rmnConstructorArgs() internal pure returns (RMN.Config memory) {
-    RMN.Voter[] memory voters = new RMN.Voter[](4);
-    voters[0] = RMN.Voter({
-      blessVoteAddr: BLESS_VOTER_1,
-      curseVoteAddr: CURSE_VOTER_1,
-      curseUnvoteAddr: CURSE_UNVOTER_1,
-      blessWeight: WEIGHT_1,
-      curseWeight: WEIGHT_1
-    });
-    voters[1] = RMN.Voter({
-      blessVoteAddr: BLESS_VOTER_2,
-      curseVoteAddr: CURSE_VOTER_2,
-      curseUnvoteAddr: CURSE_UNVOTER_2,
-      blessWeight: WEIGHT_10,
-      curseWeight: WEIGHT_10
-    });
-    voters[2] = RMN.Voter({
-      blessVoteAddr: BLESS_VOTER_3,
-      curseVoteAddr: CURSE_VOTER_3,
-      curseUnvoteAddr: CURSE_UNVOTER_3,
-      blessWeight: WEIGHT_20,
-      curseWeight: WEIGHT_20
-    });
-    voters[3] = RMN.Voter({
-      blessVoteAddr: BLESS_VOTER_4,
-      curseVoteAddr: CURSE_VOTER_4,
-      curseUnvoteAddr: CURSE_UNVOTER_4,
-      blessWeight: WEIGHT_40,
-      curseWeight: WEIGHT_40
-    });
-    return RMN.Config({
-      voters: voters,
-      blessWeightThreshold: WEIGHT_10 + WEIGHT_20 + WEIGHT_40,
-      curseWeightThreshold: WEIGHT_1 + WEIGHT_10 + WEIGHT_20 + WEIGHT_40
-    });
-  }
-
-  uint8 internal constant ZERO = 0;
-  uint8 internal constant WEIGHT_1 = 1;
-  uint8 internal constant WEIGHT_10 = 10;
-  uint8 internal constant WEIGHT_20 = 20;
-  uint8 internal constant WEIGHT_40 = 40;
 
   // Message info
   uint64 internal constant SOURCE_CHAIN_SELECTOR = 1;
@@ -116,88 +53,12 @@ contract StructFactory {
   uint16 internal constant DEST_GAS_DATA_AVAILABILITY_MULTIPLIER_BPS = 6840;
 
   // OffRamp
-  uint256 internal constant POOL_BALANCE = 1e25;
-  uint32 internal constant EXECUTION_DELAY_SECONDS = 0;
   uint32 internal constant MAX_DATA_SIZE = 30_000;
   uint16 internal constant MAX_TOKENS_LENGTH = 5;
   uint32 internal constant MAX_TOKEN_POOL_RELEASE_OR_MINT_GAS = 200_000;
   uint16 internal constant GAS_FOR_CALL_EXACT_CHECK = 5000;
   uint32 internal constant PERMISSION_LESS_EXECUTION_THRESHOLD_SECONDS = 500;
   uint32 internal constant MAX_GAS_LIMIT = 4_000_000;
-
-  function generateManualGasLimit(uint256 callDataLength) internal view returns (uint256) {
-    return ((gasleft() - 2 * (16 * callDataLength + GAS_FOR_CALL_EXACT_CHECK)) * 62) / 64;
-  }
-
-  function generateDynamicOffRampConfig(
-    address router,
-    address priceRegistry
-  ) internal pure returns (EVM2EVMOffRamp.DynamicConfig memory) {
-    return EVM2EVMOffRamp.DynamicConfig({
-      permissionLessExecutionThresholdSeconds: PERMISSION_LESS_EXECUTION_THRESHOLD_SECONDS,
-      router: router,
-      priceRegistry: priceRegistry,
-      maxNumberOfTokensPerMsg: MAX_TOKENS_LENGTH,
-      maxDataBytes: MAX_DATA_SIZE,
-      maxPoolReleaseOrMintGas: MAX_TOKEN_POOL_RELEASE_OR_MINT_GAS
-    });
-  }
-
-  function generateDynamicMultiOffRampConfig(
-    address router,
-    address priceRegistry
-  ) internal pure returns (EVM2EVMMultiOffRamp.DynamicConfig memory) {
-    return EVM2EVMMultiOffRamp.DynamicConfig({
-      permissionLessExecutionThresholdSeconds: PERMISSION_LESS_EXECUTION_THRESHOLD_SECONDS,
-      router: router,
-      priceRegistry: priceRegistry,
-      maxNumberOfTokensPerMsg: MAX_TOKENS_LENGTH,
-      maxDataBytes: MAX_DATA_SIZE,
-      maxPoolReleaseOrMintGas: MAX_TOKEN_POOL_RELEASE_OR_MINT_GAS
-    });
-  }
-
-  function generateDynamicOnRampConfig(
-    address router,
-    address priceRegistry,
-    address tokenAdminRegistry
-  ) internal pure returns (EVM2EVMOnRamp.DynamicConfig memory) {
-    return EVM2EVMOnRamp.DynamicConfig({
-      router: router,
-      maxNumberOfTokensPerMsg: MAX_TOKENS_LENGTH,
-      destGasOverhead: DEST_GAS_OVERHEAD,
-      destGasPerPayloadByte: DEST_GAS_PER_PAYLOAD_BYTE,
-      destDataAvailabilityOverheadGas: DEST_DATA_AVAILABILITY_OVERHEAD_GAS,
-      destGasPerDataAvailabilityByte: DEST_GAS_PER_DATA_AVAILABILITY_BYTE,
-      destDataAvailabilityMultiplierBps: DEST_GAS_DATA_AVAILABILITY_MULTIPLIER_BPS,
-      priceRegistry: priceRegistry,
-      maxDataBytes: MAX_DATA_SIZE,
-      maxPerMsgGasLimit: MAX_GAS_LIMIT,
-      tokenAdminRegistry: tokenAdminRegistry,
-      defaultTokenFeeUSDCents: DEFAULT_TOKEN_FEE_USD_CENTS,
-      defaultTokenDestGasOverhead: DEFAULT_TOKEN_DEST_GAS_OVERHEAD,
-      defaultTokenDestBytesOverhead: DEFAULT_TOKEN_BYTES_OVERHEAD
-    });
-  }
-
-  function getTokensAndPools(
-    address[] memory sourceTokens,
-    IPool[] memory pools
-  ) internal pure returns (Internal.PoolUpdate[] memory) {
-    Internal.PoolUpdate[] memory tokensAndPools = new Internal.PoolUpdate[](sourceTokens.length);
-    for (uint256 i = 0; i < sourceTokens.length; ++i) {
-      tokensAndPools[i] = Internal.PoolUpdate({token: sourceTokens[i], pool: address(pools[i])});
-    }
-    return tokensAndPools;
-  }
-
-  function getNopsAndWeights() internal pure returns (EVM2EVMOnRamp.NopAndWeight[] memory) {
-    EVM2EVMOnRamp.NopAndWeight[] memory nopsAndWeights = new EVM2EVMOnRamp.NopAndWeight[](3);
-    nopsAndWeights[0] = EVM2EVMOnRamp.NopAndWeight({nop: USER_1, weight: 19284});
-    nopsAndWeights[1] = EVM2EVMOnRamp.NopAndWeight({nop: USER_2, weight: 52935});
-    nopsAndWeights[2] = EVM2EVMOnRamp.NopAndWeight({nop: USER_3, weight: 8});
-    return nopsAndWeights;
-  }
 
   // Rate limiter
   address internal constant ADMIN = 0x11118e64e1FB0c487f25dD6D3601FF6aF8d32E4e;
@@ -234,51 +95,5 @@ contract StructFactory {
       Internal.PriceUpdates({tokenPriceUpdates: new Internal.TokenPriceUpdate[](0), gasPriceUpdates: gasPriceUpdates});
 
     return priceUpdates;
-  }
-
-  function getSingleTokenAndGasPriceUpdateStruct(
-    address token,
-    uint224 price,
-    uint64 chainSelector,
-    uint224 usdPerUnitGas
-  ) internal pure returns (Internal.PriceUpdates memory) {
-    Internal.PriceUpdates memory update = getSingleTokenPriceUpdateStruct(token, price);
-    update.gasPriceUpdates = getSingleGasPriceUpdateStruct(chainSelector, usdPerUnitGas).gasPriceUpdates;
-    return update;
-  }
-
-  function getPriceUpdatesStruct(
-    address[] memory tokens,
-    uint224[] memory prices
-  ) internal pure returns (Internal.PriceUpdates memory) {
-    uint256 length = tokens.length;
-
-    Internal.TokenPriceUpdate[] memory tokenPriceUpdates = new Internal.TokenPriceUpdate[](length);
-    for (uint256 i = 0; i < length; ++i) {
-      tokenPriceUpdates[i] = Internal.TokenPriceUpdate({sourceToken: tokens[i], usdPerToken: prices[i]});
-    }
-    Internal.PriceUpdates memory priceUpdates =
-      Internal.PriceUpdates({tokenPriceUpdates: tokenPriceUpdates, gasPriceUpdates: new Internal.GasPriceUpdate[](0)});
-
-    return priceUpdates;
-  }
-
-  function getSingleTokenPriceFeedUpdateStruct(
-    address sourceToken,
-    address dataFeedAddress,
-    uint8 tokenDecimals
-  ) internal pure returns (PriceRegistry.TokenPriceFeedUpdate memory) {
-    return PriceRegistry.TokenPriceFeedUpdate({
-      sourceToken: sourceToken,
-      feedConfig: IPriceRegistry.TokenPriceFeedConfig({dataFeedAddress: dataFeedAddress, tokenDecimals: tokenDecimals})
-    });
-  }
-
-  // OffRamp
-  function getEmptyPriceUpdates() internal pure returns (Internal.PriceUpdates memory priceUpdates) {
-    return Internal.PriceUpdates({
-      tokenPriceUpdates: new Internal.TokenPriceUpdate[](0),
-      gasPriceUpdates: new Internal.GasPriceUpdate[](0)
-    });
   }
 }