--- a/contracts/src/v0.8/ccip/test/TokenSetup.t.sol
+++ b/contracts/src/v0.8/ccip/test/TokenSetup.t.sol
@@ -5,7 +5,6 @@ import {IPool} from "../interfaces/IPool.sol";
 
 import {BurnMintERC677} from "../../shared/token/ERC677/BurnMintERC677.sol";
 import {Client} from "../libraries/Client.sol";
-import {RateLimiter} from "../libraries/RateLimiter.sol";
 import {BurnMintTokenPool} from "../pools/BurnMintTokenPool.sol";
 import {LockReleaseTokenPool} from "../pools/LockReleaseTokenPool.sol";
 import {TokenPool} from "../pools/TokenPool.sol";
@@ -13,7 +12,6 @@ import {TokenAdminRegistry} from "../tokenAdminRegistry/TokenAdminRegistry.sol";
 import {MaybeRevertingBurnMintTokenPool} from "./helpers/MaybeRevertingBurnMintTokenPool.sol";
 import {RouterSetup} from "./router/RouterSetup.t.sol";
 
-import {MockV3Aggregator} from "../../tests/MockV3Aggregator.sol";
 import {IERC20} from "../../vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/IERC20.sol";
 
 contract TokenSetup is RouterSetup {
@@ -28,7 +26,6 @@ contract TokenSetup is RouterSetup {
   mapping(address sourceToken => address sourcePool) internal s_sourcePoolByToken;
   mapping(address sourceToken => address destPool) internal s_destPoolBySourceToken;
   mapping(address destToken => address destPool) internal s_destPoolByToken;
-  mapping(address token => address dataFeedAddress) internal s_dataFeedByToken;
 
   function _deploySourceToken(string memory tokenName, uint256 dealAmount, uint8 decimals) internal returns (address) {
     BurnMintERC677 token = new BurnMintERC677(tokenName, tokenName, decimals, 0);
@@ -79,12 +76,6 @@ contract TokenSetup is RouterSetup {
     }
   }
 
-  function _deployTokenPriceDataFeed(address token, uint8 decimals, int256 initialAnswer) internal returns (address) {
-    MockV3Aggregator dataFeed = new MockV3Aggregator(decimals, initialAnswer);
-    s_dataFeedByToken[token] = address(dataFeed);
-    return address(dataFeed);
-  }
-
   function setUp() public virtual override {
     RouterSetup.setUp();
 
@@ -97,11 +88,9 @@ contract TokenSetup is RouterSetup {
     address sourceLink = _deploySourceToken("sLINK", type(uint256).max, 18);
     _deployLockReleasePool(sourceLink, true);
     s_sourceFeeToken = sourceLink;
-    _deployTokenPriceDataFeed(sourceLink, 8, 1e8);
 
     address sourceEth = _deploySourceToken("sETH", 2 ** 128, 18);
     _deployTokenAndBurnMintPool(sourceEth, true);
-    _deployTokenPriceDataFeed(sourceEth, 8, 1e11);
 
     // Destination tokens & pools
     address destLink = _deployDestToken("dLINK", type(uint256).max);