--- a/contracts/src/v0.8/ccip/test/libraries/Pool.t.sol
+++ b/contracts/src/v0.8/ccip/test/libraries/Pool.t.sol
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.0;
-
-import {Pool} from "../../libraries/Pool.sol";
-import {BaseTest} from "../BaseTest.t.sol";
-import "forge-std/console.sol";