--- a/contracts/src/v0.8/ccip/test/libraries/RateLimiter.t.sol
+++ b/contracts/src/v0.8/ccip/test/libraries/RateLimiter.t.sol
@@ -2,16 +2,16 @@
 pragma solidity 0.8.24;
 
 import {RateLimiter} from "../../libraries/RateLimiter.sol";
-import {BaseTest} from "../BaseTest.t.sol";
 import {RateLimiterHelper} from "../helpers/RateLimiterHelper.sol";
+import {Test} from "forge-std/Test.sol";
 
-contract RateLimiterSetup is BaseTest {
+contract RateLimiterSetup is Test {
   RateLimiterHelper internal s_helper;
   RateLimiter.Config internal s_config;
 
-  function setUp() public virtual override {
-    BaseTest.setUp();
+  uint256 internal constant BLOCK_TIME = 1234567890;
 
+  function setUp() public virtual {
     s_config = RateLimiter.Config({isEnabled: true, rate: 5, capacity: 100});
     s_helper = new RateLimiterHelper(s_config);
   }