… contracts/.changeset/fair-snails-grow.md
… contracts/gas-snapshots/ccip.gas-snapshot
… contracts/package.json
… contracts/src/v0.8/ccip/test/offRamp/EVM2EVMMultiOffRampSetup.t.sol
… contracts/src/v0.8/ccip/test/offRamp/EVM2EVMOffRampSetup.t.sol
… contracts/src/v0.8/ccip/test/onRamp/EVM2EVMOnRampSetup.t.sol
… contracts/src/v0.8/ccip/test/pools/USDCTokenPool.t.sol
… contracts/src/v0.8/ccip/test/priceRegistry/PriceRegistry.t.sol
… contracts/src/v0.8/ccip/test/rateLimiter/MultiAggregateRateLimiter.t.sol
… contracts/src/v0.8/ccip/test/router/Router.t.sol
… contracts/src/v0.8/ccip/test/tokenAdminRegistry/RegistryModuleOwnerCustom.t.sol
… core/gethwrappers/ccip/generated/mock_arm_contract/mock_arm_contract.go
… core/gethwrappers/ccip/generated/mock_usdc_token_messenger/mock_usdc_token_messenger.go
… core/gethwrappers/ccip/generation/generated-wrapper-dependency-versions-do-not-edit.txt