## Motivation
Compile times for any change to any contract have gotten out of hand. This meant that changing a single line in the on/offRamp would lead to a compile time of almost 3 minutes (M1 Pro). This made running tests to validate changes extremely slow.

## Solution
The test setups have been refactored and functionality has been moved up as far as possible. Some contracts have been able to use the forge-std base test over the shared baseTest.sol. This has lead to the following performance improvements

|         | Old time | New time | Files compiled old | Files compiled new |
|---------|----------|----------|--------------------|--------------------|
| OnRamp  | 173s     | 98s      | 44                 | 18                 |
| OffRamp | 166s     | 75s      | 44                 | 10                 |